website/backend/utils.py
import cv2
import numpy as np
from PIL import Image
import io
from typing import Tuple, Dict, Any
import logging

logger = logging.getLogger(__name__)

def validate_image(image_bytes: bytes) -> Tuple[bool, str, Dict[str, Any]]:
    try:
        image = Image.open(io.BytesIO(image_bytes))
        
        width, height = image.size
        format = image.format
        mode = image.mode
        
        if width < 128 or height < 128:
            return False, "Image resolution too low (minimum 128x128)", {}
        
        if width > 4096 or height > 4096:
            return False, "Image resolution too high (maximum 4096x4096)", {}
        
        file_size_mb = len(image_bytes) / (1024 * 1024)
        if file_size_mb > 10:
            return False, "File size too large (maximum 10MB)", {}
        
        image_info = {
            "width": width,
            "height": height,
            "format": format,
            "mode": mode,
            "file_size_mb": round(file_size_mb, 2)
        }
        
        return True, "", image_info
        
    except Exception as e:
        logger.error(f"Image validation error: {str(e)}")
        return False, "Invalid image file", {}

def preprocess_image(image_bytes: bytes) -> np.ndarray:
    try:
        image = Image.open(io.BytesIO(image_bytes))
        
        if image.mode != 'RGB':
            image = image.convert('RGB')
        
        image = image.resize((224, 224), Image.Resampling.LANCZOS)

        img_array = np.array(image).astype(np.float32)
        
        img_array = img_array / 255.0
        
        mean = np.array([0.485, 0.456, 0.406])
        std = np.array([0.229, 0.224, 0.225])
        img_array = (img_array - mean) / std
        
        img_array = np.transpose(img_array, (2, 0, 1))
        
        img_array = np.expand_dims(img_array, axis=0)
        
        return img_array.astype(np.float32)
        
    except Exception as e:
        logger.error(f"Image preprocessing error: {str(e)}")
        raise

def generate_attention_heatmap(image: np.ndarray, attention_weights: np.ndarray) -> np.ndarray:

    h, w = image.shape[:2]
    heatmap = np.random.rand(h, w)
    return heatmap