website/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
onnxruntime-gpu==1.16.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.0
redis==5.0.0
torch==2.0.1